requirements.txt
# Core dependencies
pandas>=2.0.0
pyarrow>=14.0.0
duckdb>=0.9.0

# Text processing
rapidfuzz>=3.5.0
tldextract>=5.0.0
phonenumbers>=8.13.0
unidecode>=1.3.0

# Machine learning
scikit-learn>=1.3.0
numpy>=1.24.0

# Deep learning / embeddings
torch>=2.0.0
sentence-transformers>=2.2.0

# FAISS (GPU version for Colab)
faiss-gpu>=1.7.0; sys_platform == 'linux'
faiss-cpu>=1.7.0; sys_platform != 'linux'

# Graph operations
networkx>=3.0

# Utilities
pyyaml>=6.0
orjson>=3.9.0
tqdm>=4.65.0
joblib>=1.3.0

# Country/geo handling
country-converter>=1.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Notebooks
jupyter>=1.0.0
ipywidgets>=8.0.0
